1) Работа с Javafx: - подключила к Идее, через плагин не удавалось добиться отображения элементов на окне, ошибку найти не смогла - набросала первичный корявый интерфейс - перенесла работу в контроллер - начала делать заготовки на авторизацию
