+++ client/src/main/resources/window.fxml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.ListView?>
+
+<VBox alignment="center" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8" fx:controller="ru.gb.storage.client.Controller">
+   <children>
+       <HBox fx:id="loginPanel">
+          <children>
+              <TextField fx:id="loginField" onAction="#login" prefHeight="31.0" prefWidth="204.0" promptText="Введите имя пользователя" HBox.hgrow="ALWAYS" />
+              <PasswordField fx:id="passwordField" onAction="#login" promptText="Введите пароль" HBox.hgrow="ALWAYS" />
+              <Button onAction="#login" prefHeight="31.0" prefWidth="165.0" text="Зарегистрироваться" />
+              <Button onAction="#login" text="Авторизоваться" />
+          </children>
+       </HBox>
+       <HBox prefHeight="431.0" prefWidth="660.0">
+          <children>
+              <VBox fx:id="clientStoragePanel">
+                 <children>
+                     <HBox fx:id="clientButtonPanel">
+                        <children>
+                            <Button text="Отправить" />
+                        </children>
+                     </HBox>
+                     <ListView fx:id="clientsFilesList" prefHeight="400.0" prefWidth="330.0" />
+                 </children>
+      
+              </VBox>
+              <VBox fx:id="serverStoragePanel" alignment="TOP_RIGHT" prefWidth="300">
+                 <children>
+                     <HBox fx:id="serverButtonPanel">
+                        <children>
+                            <Button text="Получить" />
+                            <Button text="Удалить" />
+                        </children>
+                     </HBox>
+                     <ListView fx:id="ServerFilesList" maxWidth="533.0" minWidth="300" prefHeight="400.0" prefWidth="330.0" />
+                 </children>
+              </VBox>
+          </children>
+       </HBox>
+   </children>
+</VBox>
